Better Process exceptions, assembly info & more
