+++ WPMTK/Properties/AssemblyInfo.cs
@@ -8,9 +8,9 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("WPMTK")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
+[assembly: AssemblyCompany("sistemiinferno")]
 [assembly: AssemblyProduct("WPMTK")]
-[assembly: AssemblyCopyright("Copyright ©  2017")]
+[assembly: AssemblyCopyright("Copyright © sistemiinferno 2017")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
